drivers: modem: gsm: move header file into modem directory
There is a specified location for modem drivers so move this
header right there.
